pyproject.toml
[project]
name = "native-blend-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "python-multipart>=0.0.22",
    "typer-cli>=0.24.0",
    "uvicorn>=0.41.0",
]
